initial typescript conversion
- install needed deps
- change files from .js to .ts
- prettier
