+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+    "trailingComma": "none",
+    "tabWidth": 2,
+    "semi": true,
+    "singleQuote": false
+}
