+++ package.json
@@ -1,32 +1,35 @@
 {
-  "name": "nodetl",
-  "version": "1.0.0",
-  "description": "A NodeJS ETL pipeline for BigQuery",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@google-cloud/bigquery": "^5.11.0",
-    "@google-cloud/storage": "^5.18.2",
-    "axios": "^0.26.1"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
-    "eslint": "^8.10.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^16.1.1",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.3",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "typescript": "^4.6.2"
-  },
-  "imports": {
-    "#src/*": "./src/*"
-  }
+    "name": "nodetl",
+    "version": "1.0.0",
+    "description": "A NodeJS ETL pipeline for BigQuery",
+    "main": "dist/index.js",
+    "scripts": {
+        "start": "node dist/index.js",
+        "dev": "ts-node src/index.ts",
+        "build": "tsc"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@google-cloud/bigquery": "^5.11.0",
+        "@google-cloud/storage": "^5.18.2",
+        "axios": "^0.26.1"
+    },
+    "devDependencies": {
+        "@typescript-eslint/eslint-plugin": "^5.14.0",
+        "@typescript-eslint/parser": "^5.14.0",
+        "eslint": "^8.10.0",
+        "eslint-config-airbnb": "^19.0.4",
+        "eslint-config-airbnb-typescript": "^16.1.1",
+        "eslint-plugin-import": "^2.25.4",
+        "eslint-plugin-jsx-a11y": "^6.5.1",
+        "eslint-plugin-react": "^7.29.3",
+        "eslint-plugin-react-hooks": "^4.3.0",
+        "ts-node": "^10.7.0",
+        "typescript": "^4.6.2"
+    },
+    "imports": {
+        "#src/*": "./src/*"
+    }
 }
