+++ src/extractors/RandomUserExtractor.ts
@@ -1,11 +1,11 @@
-const AbstractExtractor = require('./AbstractExtractor');
+const AbstractExtractor = require("./AbstractExtractor");
 
 class RandomUserExtractor extends AbstractExtractor {
   constructor() {
     super();
 
-    this.apiBase = 'https://random-data-api.com/api/';
-    this.endpoint = 'users/random_user';
+    this.apiBase = "https://random-data-api.com/api/";
+    this.endpoint = "users/random_user";
   }
 
   extractData(params = {}) {
