+++ src/transformers/AbstractTransformer.ts
@@ -1,5 +1,5 @@
 class AbstractTransformer {
-  key = '';
+  key = "";
 
   /**
    * @param {{}} data
